Refactor LAS constants, utils into a subdirectory
