+++ lit/src/shared/constants.ts
@@ -16,9 +16,6 @@ export const STAGE_DESCRIPTION_VOTE = "Vote for a participant leader";
 export const STAGE_DESCRIPTION_REVEAL =
   "Show stage results. This stage is implicitly added based on other stages. To delete this stage, adjust reveal settings in relevant stages.";
 
-export const GAME_DESCRIPTION_LAS =
-  "An implementation of the Lost at Sea task (Born 2022) with pairwise elements.";
-
 /** Profile avatars. */
 export const PROFILE_AVATARS = [
   '👩🏻','👩🏼','👩🏽','👩🏾','👩🏿',
